Misc update; Ecomm website misc FE update; Progress

+++ resources/views/livewire/ecomm-website/home-component.blade.php
@@ -76,39 +76,41 @@
 
         <div class="row">
 
-          @for ($ii=1; $ii <= \App\Product::count(); $ii++)
-            <div class="col-md-3">
-              <div class="card h-100 shadow">
+          @for ($ii=1; $ii <= \App\Product::max('product_id') && $ii <= 4; $ii++)
+            @if (\App\Product::find($ii))
+              <div class="col-md-3">
+                <div class="card h-100 shadow">
       
-                <div class="d-flex flex-column justify-content-between h-100 bg-success-rm">
-                  <a href="{{ route('website-product-view', [\App\Product::find($ii)->product_id, \App\Product::find($ii)->name]) }}">
-                    <div class="d-flex justify-content-center bg-warning-rm">
-                        @if ($productCategory->image_path)
-                          <img class="img-fluid h-25-rm w-100-rm" src="{{ asset('storage/' . \App\Product::find($ii)->image_path) }}" alt="{{ $productCategory->name }}" style="max-height: 150px; {{--max-width: 100px;--}}">
-                        @else
-                          <i class="fas fa-dice-d6 fa-8x text-muted m-5"></i>
-                        @endif
-                    </div>
-                  </a>
+                  <div class="d-flex flex-column justify-content-between h-100 bg-success-rm">
+                    <a href="{{ route('website-product-view', [\App\Product::find($ii)->product_id, \App\Product::find($ii)->name]) }}">
+                      <div class="d-flex justify-content-center bg-warning-rm">
+                          @if ($productCategory->image_path)
+                            <img class="img-fluid h-25-rm w-100-rm" src="{{ asset('storage/' . \App\Product::find($ii)->image_path) }}" alt="{{ $productCategory->name }}" style="max-height: 150px; {{--max-width: 100px;--}}">
+                          @else
+                            <i class="fas fa-dice-d6 fa-8x text-muted m-5"></i>
+                          @endif
+                      </div>
+                    </a>
       
-                  <div class="d-flex flex-column justify-content-between flex-grow-1 overflow-auto" style="background-color: #f5f5f5;">
-                    <div class="p-2">
-                      <a href="{{ route('website-product-view', [\App\Product::find($ii)->product_id, \App\Product::find($ii)->name]) }}">
-                        <h2 class="h4 font-weight-bold mt-2 mb-2 text-dark" style="font-family: Arial;">
-                          {{ strtoupper(\App\Product::find($ii)->name) }}
-                          <br />
-                          <span class="text-danger ml-1-rm">
-                            Rs {{ \App\Product::find($ii)->selling_price }}
-                          </span>
-                        </h2>
-                      </a>
+                    <div class="d-flex flex-column justify-content-between flex-grow-1 overflow-auto" style="background-color: #f5f5f5;">
+                      <div class="p-2">
+                        <a href="{{ route('website-product-view', [\App\Product::find($ii)->product_id, \App\Product::find($ii)->name]) }}">
+                          <h2 class="h4 font-weight-bold mt-2 mb-2 text-dark" style="font-family: Arial;">
+                            {{ strtoupper(\App\Product::find($ii)->name) }}
+                            <br />
+                            <span class="text-danger ml-1-rm">
+                              Rs {{ \App\Product::find($ii)->selling_price }}
+                            </span>
+                          </h2>
+                        </a>
       
+                      </div>
                     </div>
                   </div>
-                </div>
 
+                </div>
               </div>
-            </div>
+            @endif
           @endfor
 
         </div>
